Fix #265 - Colour change escape sequences in Label() strings don't work correctly
Use correct interpretation of hexadecimal characters to integers.

Also fixed a bug where if the string was shorter than 8 bytes, the
returned value was 16 times larger than it should be (formerly
"123456" was interpreted as 0x01234560).

Non-hex characters are now consistently interpreted as '0'.

The fix has been compiled and tested in isolation, but I can't
compile the modified gpixmap.h in my environment.