Possibility to set custom pid for parsing epg data
This is very useful if a transponder epg be filed in 2 or more different pids or use non stardart one.
For example, the platform Dolce sent epg data on Romanian language on PID 0x12 and Bulgarian and Serbian on PID 0x12b. If create reader for PID 0x12b, some events are Romanian language, others are Bulgarian because of PID 0x12 no information about the language of epg data.
Usage:
Create a file "/etc/enigma2/epgpids.cusom" and fill in the format

Orbital position | TSID | ONID | Custom pid to use

The values ​​can be decimal or hexadecimal (0x ...), and combined
For orbital position using integers where 1 means 0,1 ° (19,2°E -> 192)
For orbital position can be used as negative numbers (0,8°W -> -8 or 3592)

examples:
390 17 28 299 # 39°E 12688V, epgpid to use: 299
0x186 0x11 0x1c 0x12b # Same, but in hexadecimal format
390 17 28 0x12b # Or combined