[frontend] fix incorrect init and usage of LOF offset
The local oscillator frequency offset is incorrectly initialized to -1.

Also when enigma was requesting the frequency, offset value was added for systems other than satellite.
This was causing changes to frequency by -1 Hz. Not a big deal, but when using a multituner device
that supports dvb-s/c/t then changing from s to c/t offset (eg 9750000-1) was added in frequency causing tune failure.

This patch initializes FREQ_OFFSET to correct value 0 and takes into consideration system to be satellite in order to add up offset.

Credit: Athoik