Add graphical switch icon option for Boolean settings
This change adds the infrastructure for a skin to enable a graphical switch icon to replace the On/Off, Yes/No, True/False and Enabled/Disabled configuration selection options.

To be properly implements the switch images and a configuration block must be added to the active skin.  I propose to add the images to the default skin so that they will be available to all skins.  I also would like to add the appropriate block to the default skin.  Other skin maintainers can add the block and, if they desire, alternate images.