[Transponder.py] Improve DVB-T bandwidth logic
This improvement matches Fix #147 where the bandwidth is now reported in Hz.

In this case the improvement removes a potential limitation where the bandwidth does not match any of the entries in the current fixed list.  The actual bandwidth is reported directly using the same format as was listed in the fixed table.