Start / End bookmarks missing on recordings after being chaseplayed
The start and end bookmarks can be lost in an active recording if
the recording is played back via the media player while the recording
is still happening.

Replication: Start a recording, and while the recording is still
in progress, start playback in the MoviePlayer.

When the recording finishes, stop playback or allow playback to
end.

Play the recording again. The start/end bookmarks will be missing.

This fix preserves the bookmarks and has the additional benefit of
displaying "start/end" bookmarks in chaseplay as soon as the bookmarks
are generated in the corresponding recording object.

[InfoBarGenerics]

If a recording is played while it is still being recorded, identify
the corresponding eDVBServiceRecord object, and when it issues
evNewEventInfo events, update the InfoBarCueSheetSupport.cut_list
and playback's eDVBServicePlay object's cut list.

This will ensure that the playback has the recording's "start/end"
bookmarks available to write to the .cuts file when the playback
is stopped.

If the recording is still being made when playback is finished,
remove the recording's "start/end" marks up to that point from the
eDVBServicePlay object's cut list. The corresponding eDVBServiceRecord
object will write these "start/end" marks, and any that are defined
later in the recording, when the recording finishes.

If the recording is played once more while it is still being recorded,
the current "start/end" marks will be retrieved from the from the
eDVBServicePlay object's cut list.

Because the eDVBServicePlay object's cut list is updated in real
time, the "start/end" marks appear dynamically in the media playback's
infobar.

[iRecordableService,eDVBServiceRecord]

Add new method getCutList() that returns a list of the "start/end"
bookmarks in a recording object.

A default iRecordableService::getCutList(), which returns an empty
cut list, is defined so that other classes that derive from
iRecordableService need not define getCutList().

Extract common code in eDVBServiceRecord::saveCutlist() and
eDVBServiceRecord::getCutList() into private member function
eDVBServiceRecord::fixupCuts().

Remove comment about refactoring eDVBServicePlay::saveCuesheet()
and eDVBServiceRecord::saveCutlist(), because they are really only
superficially similar, and don't share much code.

[eDVBServicePlay]

In stop(), test whether the cuesheen needs to be saves after stop()
issues an evStopped event. This is required because new code in
Screens.InfoBarGenerics.InfoBarCueSheetSupport may call setCutList()
during in code called by the evStopped event.