Sort timezones on "user-friendly" name
Except for the "GMT" timezone names, sort timezone names on the
user-friendly name rather than the system time zone name. This
presents the lists sorted on what the user sees in the ConfigSelection
rather than on the system timezone names.

The new sort is implemented in sorttzChoices(). For backwards
compatibility, the old function sorttz() is re-implemented by calling
sorttzChoices().

I've added "Brazil/DeNoronha" expanded as "Fernando de Noronha" as
an example. Its sort order is now correctly after "Brazil/East"
rather than before it, if it were sorted on "DeNoronha".

It would also make sorting work correctly if translations of place
names are used (e.g. Cologne -> Koeln, Vienna -> Wien).

OpenViX devels may want to look at other possible name expansions
in the OpenViX zoneinfo database, which is more extensive than the
Beyonwiz one. In particular, '-' seems to be used for a number of
different purposes: as a real part of the location name, as in
America/Port-au-Prince, as an alternative to "_" (e.g.
Canada/East-Saskatchewan) and as a hierarchy separator, as in
US/Indiana-Starke (for, I presume, Starke County, Indiana).