[ChannelSelection] Use fallback if no network time and not on DVB service (#183)
If the initial live TV service at startup isn't a DVB service
(more accurately, isn't a service that supports the Time Of Day table)
and the system has no internet connection, the system time will
remain at the system startup default of 00:00:00 1 Jan 1970. An
example of such a non-broadcast service is the HDMI IN service on the
Beyonwiz T4.

This will mean that the displayed time on the PVR will be wrong,
any recordings that are due won't be made, an the default view of
the EPG will be for 1 Jan 1970.

Replication steps

If possible, select a live TV service that is not a DVB broadcast,
disconnect the PVR from the network and reboot.

If that isn't feasible, in the command-line interface:

init 4

Edit /etc/enigma2/settings so that config.tv.lastservice is a valid
service ref, but not a DVB or similar service (i.e. has on Time Of
Day table), e.g. 8192:0:1:0:0:0:0:0:0:0:. For that particular
example, it doesn't matter whether the device supports HDMI in.

Then:

halt

Then disconnect the PVR from the network and restart.

In either case, the time will start from 00:00:00 1 Jan 1970 when
the system restarts, and continue from that time setting until a
valid DVB channel is selected.

Fix:

If at startup config.tv.lastservice is not valid or, if it is valid,
but not a DVB service and the time is < 1 Jan 2004, find the first
valid DVB service in the current bouquet, or if that fails and
config.usage.multibouquet is set, then the first valid DVB service
in any other bouquet.

As well as fixing this bug, the fix will try to find a valid DVB
service if config.tv.lastservice is not valid.

The new methods ChannelSelection::findServiceBouquet() and
ChannelSelection::forceZap() are intended to be able to be called
to replace most of the code in RecordTimer::failureCB(). They may also
be able to be adapted to replace most of the code involved in
zapping by channel number (and the channel panic button action).