[scan] do not ingore namespace subnetwork By default the namespace contains only part of the hash that is possible to store because when we are having valid ONID we are ingoring the last 2 bytes (subnetwork).
That is happening when isValidONIDTSID is true and that exists since the initial
import of enigma2 in git.

In the subnetwork we are storing the frequency and polarization for sat
and frequency for other systems and that solves all the problematic cases where
same ONID is appeared twice, without hardcoding exceptions in isValidONIDTSID.

Although by not ingoring the subnetwork all the service references will change
making at least picons unusable.

The findPicon can be optimized to ingore the subnetwork and that will be fixed
by another.

Ideally the namespace should not ingore the subnetwork and isValidONIDTSID
check can go away, once there are no issues preventing to do so.