Revert "Add fast network availability test"
Requested by Andy

SoftwareUpdate/PluginBrowser/OnlineUpdateCheck: Add fast network availability test...

Add a fast (typically 0.1 sec) multithreaded simultaneous ping test of multiple hosts to check for network availability. As reported, currently we have to wait up to 30 seconds with the spinner up if there is no network availability and at the end of this time no message is sent to the GUI advising the user of the problem even though these messages are already present in the current code and printed as debug. Also add an extra MessageBox so the error messages are displayed in the SoftwareUpdate GUI on '403' and '404' responses. (reverse-merged from commit e0546964f3e8b9e0033d1bec80ba12ca95866d72)