service: Allow a plain URL to construct eServiceReference
Allow an unencoded URL to be passed to eServiceReference. If it starts with a
letter, try interpreting it as URL, hence it must start with "http://" or
something similar. The URL must NOT be encoded in that case. It may have a
name separated by a space (and also not encoded), for example:
http://myserver.local/stream.mp3 My own server

This makes adding network streams to a settings file much simpler, and a future
"play this URL" feature is also simplified.

This also shows there's absolutely no reason for the eServiceReference to require
such a weird formatting of services.