[NetworkSetup.py] Refactor NetworkPassword class
This change makes NetworkPassword use the existing Setup class rather than duplicating some of the code from within the Setup class.  The configuration list is now stored in the setup.xml file like all other standardised Setup screens.

The new code uses the standard Setup way of showing the text-style keyboard as well as the TEXT button to bring up the virtual keyboard to enter the password. This new code still offers access to the YELLOW button to select a random password but because the new code uses the standard Setup code this button will not be shown on skins that do not have a Setup screen that supports the display of all four colour buttons.

If the current skin does not support the four colour buttons then a new screen named "NetworkPassword" can be created to display the complete screen. The "NetworkPassword" screen should have all the elements of a standard "Setup" screen but should also include the YELLOW button. The "NetworkPassword" screen can also be used by skin authors who would like this screen to be differentiated from the standard "Setup" screen.

The old screen 'NetworkSetPassword' is no longer being used and can be deleted from the skins.

NOTE: I recommend that all skins should be updated to support the new button proposals in /doc/BUTTONGUIDE.  This will allow more efficient skin development and usage moving forward.

Users should be reminded that while blank passwords can be used for the network password they are VERY INSECURE and are NOT recommended for devices connected to the Internet (i.e. if a port is forwarded to the box).