Revert "Refactor NetworkPassword class (#219)"
This reverts commit c76be6aee997ddafabe0418ade262f1ea5d0de3c.

This commit hijacks setup.xml for non-menu items and will also cause merge problems. Unnecessarily  increasing the size of setup.xml is going to cause a performance hit every time the menu is opened. This will get worse as more items are added. This is not a road we want to go down.