Setup: Rename a variable
This variable name is confusing. It has "config" in the name but the reality is it could be used for any conditional statement. e.g... conditional="1+1==2" would also work in this context.

The variable name used is"conditional" the same as used in skin.py when getting conditions from the .xml file.