Fix HelpWindow position (#1472)
Add functions hideHelp and showHelp in config and use for HelpWindow.
This allow hide and show help_window in ConfigList.
Before that, the onDeselect and onSelect functions were used for it, which delete or add help_window instance.
It breaks the help_window position because it is set only once after the first help_window instance is added.
https://forums.openpli.org/topic/59254-do-commit-comments-get-read/
Conflicts:
	lib/python/Components/ConfigList.py