[ActionMap.py] Add multiple contexts to HelpableActionMaps (#1538)
Allow HelpableActionMap and HelpableNumberActionMap to accept a list of contexts as well as a single context.  This makes it possible to convert ActionMap and NumberActionMap objects to helpable without changing their basic structure.  Single-context helpable ActionMaps are retained for backwards compatibility.

This change was held back to allow for an issue with the CSFD plugin to be corrected.  I am told by Ims21 that the issue has now been corrected.