picload: guard against incomplete reads
The read is not guaranteed to return the bytes we requested.
Instead of assuming that read calls are sucessfull, check that we get the bytes requested.

In some cases an eDebug was added, at least to help tracing issue with bad image files.

Finally this commit solves several unsused-result warnings, but there are still many to go.