NimManager: add support for nim_socket type DVB-S2X
Some DVB-S2X capable frontends report their type to be DVB-S2X in
/proc/bus/nim_sockets.

Instead of renaming the type to DVB-S2, add S2X relations with DVB-S/S2
in the compatibility / connectability lists.

This allows the frontend type to remain unmodified, and be advertised as
'DVB-S2X'.

NOTE: this is a cosmetic-only change, the nim_socket description and
type should only be used to display frontend information in the user
interface.
There is no DVB-S2X delsys of course.