Stop checking for dual tuners for loopthroughts
As it seems to hide the scan configs in case you have a VU box that only
loops A to B instead of B to A. It seems als currently when you try to
config to loopthrough a tuner to a not connected tuner that the config
is not stored anyway as the and result the loopthrough is also not
connected.

And as this somewhere brakes something it can still be reverted.

See... https://forums.openpli.org/topic/60754-nimmanagersomethingconnected-broken-logic/page-2