ServiceInfo: display Gold PLS only when required
The PLS mode in latest enigma2 uses only Gold sequences. So it's waste of space to display this info.
Also change description in PLS code, since we use Gold only sequences.
Finally display the Input Stream ID only when it has values > 0. So it won't be displayed when it has default value -1.

Merged with conflicts...