Revert "[frontend.cpp/RecordTimer.py] test fix for "Unicable simultaneous recordings bug"."
This reverts commit 2b8a43869f92e7dd1b8728a4203211f9d4f459aa.

This is reverted because it failed testing by AbuBaniaz.

Tested as follows:
Vu+ Solo 4K, GT-Sat Unicable LNB.

Set a recording on BBC One HD (Astra 2).
Set a simultaneous recording on Channel 5 HD (Astra 2).
Put STB in standby.
Upon leaving standby while the recordings are in progress a "tune failed" message is displayed and the recordings are found to be zero length.