MovieSelection: use suburi to play two files in parallel
In https://github.com/OpenPLi/enigma2/commit/65962a39625aaf02f66ee9258181ec2d164916a6#diff-edf65e1563875233346cf1cf40ed3c76 is implemented suburi in service reference path string, which allows you to play two files in parallel.
But it is possible use only in bouquet or in plugins that create service reference with suburi in path.
This allows you to play files from a disk.
If a video file name ends with '_suburi' is checked if there is also an audio file with the same name without '_suburi' in end,
and if it exists then play this two files together.

It allows download separate Dash MP4 video and Dash MP4 Audio files from youtube and play together video and audio.