[config.py] Fix extra_args management (2) (#27)
* [config.py] Fix extra_args management (2)

There is a critical flaw in the ConfigElement notifier code with particular reference to the handling of the "extra_args" parameter. The current code tries to use a hash of the notifier method as the key to a dictionary. This is illegal! The errors have been hidden by try / except blocks rather than trying to understand or fix the underlying problem. The effect of the problem was that attempts to access the extra_args would fail with "TypeError: unhashable type: 'xyz'" and the version of the callback without the "extra_args" would be used instead. (This causes the "extra_args" information to be lost.)

Rather than using the notifier itself as the key use it's id() instead.  The protective try / except protection has been removed as the underlying issue has been resolved.

Thanks to IanB and Prl for assisting address this issue.

* [config.py] Add note about storing "extra_args"

Add a note that there can only be one "extra_args" saved per instance of any notifier.

* [config.py] Initialise and maintain notifiers as lists

* [config.py] Correct setNotifiers() / setNotifiersFinal()

Thanks to Prl for pointing out the issue with the previous code.