Update Timeshift.py
Add support for pause/play IPTV stream instead of displaying (or not, depending on settings) error message stating timeshift not supported.

This isn't full timeshift (ie seek) support with IPTV streams, rather a pause/playback.  Maybe over time, this could be incorporated if possible!

Many thanks to @danny121pt for posting original mod and the two timeshift commits on ATVs implementation of this.