[VirtualKeyBoard.py] Add option to select GREEN action
Add a calling parameter that selects how the GREEN button will be labeled.  The same parameter also defines how the green key map grid item will be displayed.

This change is based on an idea added to OpenPLi version of VirtualKeyBoard.  This implementation is more complete.

NOTE: The incomplete OpenPLi implementation uses a different calling parameter with a different methodology!