Solve issue satconfig always shown due to e0c312ce
When DVB-S is disabled it still showing all sat options (since .self.nimConfig.configMode.value will be not set to 'nothing').
Just add one tab to the 'menu option' code to  move to code to the same 'detention block' (or whatever it's called in Python) as 'self.list.append(self.configMode)'.
Thanks Beeker for reporting