NimManager: Solve possible GSODs
When the configMode choices are defined some options depend on other
tuner slots, and within the other slots the configMode may not be
defined yet. Therefore in NimManager we pre-defined these options and
then in Satconfig we 'pop' them when they are not required as in
Satconfig each tuner slot has already the configMode defined all tests
required can be done there.