[VirtualKeyBoard] Enhance
* [VirtualKeyBoard.py] Replace 'ALL' button with 'DEL'

This change removes the 'ALL' button from the keyboard grid as it now serves no purpose.  The functionality remains in the code should it be desired or useful in the future.  Its position on the keyboard layout is now used by the new 'DEL' button.  This change allows all the locale keyboards to return to their previous layouts and have the 'DEL' button in a consistent position.

* [VirtualKeyBoard.py] Add large button support

This change enhances the VirtualKeyBoard by adding large button support for all locales.  Characters and icons within the buttons can now have their horizontal and vertical alignment specified.  The change also optimises code execution by moving some of the processing to the init code rather than running at each invocation.

This change requires all skins to use the updated VirtualKeyBoard images.  The skins themselves do not need to be changed.

* [VIRTUALKEYBOARD] Update documentation

Update the documentation to match the code changes.