Fix rewind stops several seconds before start of file
When rewind reaches position 0 a "start of file" event is send
immediately which stops rewind. But there is still data in the
driver buffer which wasn't shown on the TV.
There is a mechanism to wait until driver buffer is empty when
playback or fast forward reaches end of file.
Patch now uses same mechanism for start of file.