Rework VirtualKeyBoard Shift functions
Change the name of the current Shift function to Caps Lock.  Add a new Function called Shift.  The new Shift function allows the selection of a different shift level for the next character only.  Once the selection is made the previous shift level is immediately restored.

The new Shift function has been assigned to the YELLOW button.  The Locale function that was on the YELLOW button has been moved to the TEXT button.  This was done as it is expected that the new Shift function could be used more frequently than the Locale function.

NOTE: This change should have a matching change to all skins to add the new vkey_shift.png icon image.
(cherry picked from commit 417e3066963440f94035ca753e38aedf9db772ef)