Allow EPG to be used on system bouquets
Allow the EPG to be used in the system bouquets in the channel
selevtion screen's All, Satellites and Providers menus.

[ChannelSelection]
Cleaned up some repeated code into static methods getTransponderName()
and getServiceType().

Added methods getEPGBouquetName() and getEPGBouquetList() to support
bouquet lists that also include the system bouquet lists.

In the future, it may be desirable to combine the common parts of
the functionality of getBouquetList() and getEPGBouquetList().

[EpgSelection]
In getBouquetService() allow the returned list to be sorted in line
with the bouquet's ORDER BY clause (if specified). System bouquets
aren't all ordered by bouquet.

[InfoBarGenerics]
In zapToService() preserve the self.servicelist.servicePath better.

In MultiServiceEPG(), use the new ChannelSelection method
getEPGBouquetList() instead of getBouquetList().

Requires the changes in commit d550d5c.