Simplify debug logging in enigma2.sh
Test directly on the exit status of grep -q, use sed -n and avoid
repeating greps to simplify testing debug log-related settings and
extracting the debug log file name from the settings.

Escape '.' characters where appropriate in regular expressions in
the debug log code so that literal '.' characters are matched, not
"any character".

Replace an instance of "/etc" in the debug log code with the
@sysconfdir@ template.

Change the redirection of debug logs to /dev/null from "2>" to "&>"
to ensure that both stdout and stderr are captured. This doesn't
seem to have a practical effect, but  it makes the redirection the
same as the redirection into a log file.