[ChoiceList.py] Fix text position when no icon is used
In Choicelist, there is an option to create it with an icon on the left. This could be the "expandable", "expanded", "verticalline" or "key_%s" icon.
The position of the icon and the list text are controlled by the "ChoicelistIcon" and the "ChoicelistName" skin parameters. So far, so good.

A choicelist can also be created without any icon at all, when the "ChoiceEntryComponent" function is called with "key=None".
In this case, the list text is aligned more to the right, leaving space for the non present icon. This is not pretty.

This commit adds a new skin parameter called "ChoicelistNameSingle" to control the position of the list text when the list is created with "key=None".
It also restricts the effect of the existing "ChoicelistName" parameter to only when "key" is not "None".

(cherry picked from commit 5456c1daccb6c47f07e84215da589472970f332d)