[GraphMultiEPG.py] Removed duplicate and buggy skin attributes
Removed 2 duplicate skin attributes ("ServiceNameForegroundColor" is the same as "ServiceForegroundColor" and "ServiceNameBackgroundColor" is the same as "ServiceBackgroundColor"). Also the "ServiceNameBackgroundColor" was listed twice, but controlled two different variables (once it set the "self.backColorService" and the other the "self.backColorRec"), while there is a another attribute ("ServiceBackgroundColorRecording") which controls the "self.backColorRec" at the same time! This was clearly a bug and a headache even for the initial developer who programmed it! The removed attributes were not used by the PLi skins (or any skins that I know), so no problem at all. Finallt the skin attributes were re-sorted, so they can be easier identified.

(cherry picked from commit a56ac85ff581eb9ac4a66c9f472b3532b04aa8f8)