[SkinSelector.py] Refactor and enhance code
This change makes all skins equal in format and structure.  It does this by using an appropriate "skin.xml" or "skin_display.xml" at the top level of the skin directory.  This the emergency skin and display skins need to be changed so that they are no longer special.  This has been done in a partner pull request for the emergency skin.

This code now has an associated HELP screen.

The UI now has the option to be more configurable via the skin.  The user can now be shown which skin is the Emergency skin, which skin is the image default skin and which skin is the currently selected skin.  There is some protection code to determine if the current skin is capable of displaying the new UI.  If a problem is detected the code will force an internal screen to be used.  This should give all skin authors time to update their skins to be compatible with the new code.  This protection code is intended to be removed when most skins have been updated.