Test update to make display800 skin compatible with new SkinSelector (#76)
* [Makefile.am] Move skin XML to new location

This change moves the skin display components for this display skin to its new location as part of the "skin.py" and "SkinSelector.py" refactors.

* [skin_display.xml] Use skin relative path for includes

* [skin_display_picon.xml] Use skin relative path for includes

* [skin_display.xml] Update location of template file

This change updates the new location of this skin's display template.

An alternate, an possibly better, update would be to add a copy of the template file into this skin.  This would mean that this skin is totally self complete.