[SkinSelector.py] Harden code against potential crash
This change places the regular expression that looks for the skin resolution into a try/except structure.  Some skins do not contain the resolution information, or even an "output" block.  This used to cause the code to crash.  Now if there is a problem finding the resolution a blank string is returned instead.