[skin.py] Allow include files to add core data
This change allows skin include files to add core skin elements like fonts, colours parameters, etc. 

NOTE: If this feature is to be used the include files *MUST* be included such that the core elements are defined *BEFORE* they are used.  That means that the order of the include files is important.