[skin.py] Optimise and enhance skinLoad() method
- Extract the common code adding screens to domScreens into a single method. This eliminates a nest of "if" blocks.
- This change also enhances the skinLoad() method to allow skin include files to add core skin elements like fonts, colours parameters, etc.

NOTE: If this enhanced skinLoad() feature is to be used the include files MUST be included such that the core elements are defined BEFORE they are used. That means that the order of the include files is important.
(cherry picked from commit dc3a7911f3de586c1069b606bccb43becddc4d17)