[keymap] Undefined actions
There are actions of the ViX code not defined in the keymap, for example, the number keys can define which audio track can be chosen.

example: enter with audio key in AudioSelection, now press one of the available audio tracks on the numeric keypad and directly enter that audio mode.
==============================

fix volumeMute From AudioSelection