[SkinSelector.py] Simplify the embedded skin
Remove use of the button background colours and simplify the embedded skin.  This is an embedded screen to ensure that this code works even if the current skin does not match the code.  The intention is to keep this code functioning while skin developers and maintainers update their skins to match this updated code.  In this context it was illogical to try and make the embedded skin a complex and fully featured skin replacement.

For this code to work the "key_*" colours need to be defined.  This is currently in the emergency skins.  It is intended to add all critical colours to "skin.py" to ensure that all important colours are pre-initialised (in the same was as key fonts are pre-initialised).