cablescan: fix comments added by a4b35bdc3df7a4bd5cf6dfa2e4ac59a53455a4ab
The code comment as well as the original commit message was incomplete.

What commit a4b35bdc3df7a4bd5cf6dfa2e4ac59a53455a4ab actually does, is
cause services to be ignored when the SDT claims they are on the actual
transponder, whereas the NIT has linked them to a different transponder.
SDT services which do not occur in the NIT, even if the NIT has a fully
populated servicelist, are still added.
(The commit message as well as the comment in the added code might have
wrongfully suggested such services would be ignored as well)