Multiboot rebuild- Part 1(Part2 in ViX plugin vix-core) (#495)
* Make multiboot less hardcode dependend

Let it fulle depend on what the STARTUP_... files do describe

* Use PosixSpawn to u(mount) for getMultibootslot

And move it from SystemInfo.py to MultiBoot.py

* Multiboot: Simplify some code and make it a bit more universal

* When no callback for ePopen do wait until command is finished

This also means PosixSpawn is not required anymore.

As far I tested I discovered no side effects. Please let me know when we
find any.

* Multiboot: Use Console().ePopen() instead of PosixSpawn

As it is not required anymare as Console().ePopen() without fallback
function now waits until the command is completed

* Multiboot: Find STARTUP device by simply scanning whitelisted devices

So the model parameter is out of getMultiBootStartupDevice which makes
the code more universal. I added also the device that is used for SF8008
and hopefully it also finds the device for GigaBlue

* Multiboot: Do not remount device was next slot uses same device

This can happen with directory based multiboot

* Multiboot sort imports

* change MultiBootSelectDevice to MBbootdevice

* multiboot add whole STARTUP name, and add kernel device to dict

* multiboot strip/n and pickup kernel correctly

* multiboot fix image date stamping

* multiboot fix empty slot routine

* Multiboot: Streamline code and make a bit more rebust for GSOD

* About sort imports

* About imlement new multiboot code

* Console - make OpenPli code more robust

* Menu - change multiboot restart to MultiBoot Image Selector

* Multiboot Image Selector - Add to Image

* Multiboot - reorder Enigma2 available check

* MultiBoot Image Selector use same skin name

* Multiboot - remove os. from paths

* SystemInfo - canBackupemc - pickup boot evice automatically

* multiboot - finalise for slot removal

* Multiboot - remove debug

Co-authored-by: littlesat <littlesat99@yahoo.com>
Co-authored-by: TwolDE <awhalley@waitrose.com>