[skin.py] More fixes and improvements (#148)
* [skin.py] Restore strip() method for parameters

The strip() method code was accidentally lost from the parameter parsing.  This resulted in parameter lists separated by commas *and* white space would fail.

* [skin.py] Speed improvement

Replace fileExists() with faster pathExists()

* [skin.py] Use isfile() rather than pathExists()

Using isfile() as a test for existence is more optimal because, in context, it is better to know that the path exists and is a file rather than just knowing the path exists.