Action file changes (#149)
* Revert "Fix key repeat for aarch64 bit systems"

This reverts commit 03027c2fed7ed560fe14be49fbd57b95a35b3da5.

* Revert "Add possibility to remove native key binding"

This reverts commit 2af8e5f8f097e6b22a6d7e03401cea4e9988e597.

* actionmap: cosmetics/debugs
20151218-1

* [Keymapper] revert to 20100107

Reverts changes in commits
0ce9f9017158a6eef8833998514ab85df7007368
bdfa50fcecbca0714c1c1143dd0d950404b17701

* actionmap: Implement key translations

Some remotes lack support for certian buttons or have buttons
with different key codes. This translation feature allows to
translate specific keys into a different one reducing to need
to create bloated keymaps to map several buttons to the same function.

Some remotes are very small and lack certain keys. Using a toggle
feature based on a specific button, buttons can be multi functional

The definitions are like a keymap.xml and the translations can be added
to the default keymap.xml. But to keep keymap.xml clean and to let it serve
as a standard keymap, enigma2 will look for a keytranslation.xml file.
In is just like a keymap.xml but can be used for STB specific mappings
and translations.

An example keytranslation.xml in which the number keys are mapped
to specific other keys is shown below.  The toggle="1" means that
translation will only be done after the toggle key has been pressed. The
toggle element defines the toggle button, in this case the BACKSPACE key.
key definitions without the toggle= attribute (or with a value not equal
to "1") will always be translated regardless the toggle state.

The device name is taken from the /dev/input identifications. Check your
enigma2 log to find the proper names.

<keymap>
      <translate>
              <device name="Small remote device name">
                      <toggle from="KEY_BACKSPACE"/>
                      <key from="KEY_1" to="KEY_RED" toggle="1"/>
                      <key from="KEY_2" to="KEY_GREEN" toggle="1"/>
                      <key from="KEY_3" to="KEY_YELLOW" toggle="1"/>
                      <key from="KEY_4" to="KEY_BLUE" toggle="1"/>
                      <key from="KEY_5" to="KEY_PREVIOUS" toggle="1"/>
                      <key from="KEY_6" to="KEY_NEXT" toggle="1"/>
                      <key from="KEY_7" to="KEY_REWIND" toggle="1"/>
                      <key from="KEY_8" to="KEY_STOP" toggle="1"/>
                      <key from="KEY_9" to="KEY_FASTFORWARD" toggle="1"/>
                      <key from="KEY_0" to="KEY_PLAYPAUSE" toggle="1"/>

                      <key from="KEY_F7" to="KEY_MENU"/>
                      <key from="KEY_F1" to="KEY_VIDEO"/>
                      <key from="KEY_HOME" to="KEY_INFO"/>
                      <key from="KEY_BACK" to="KEY_EXIT"/>
                      <key from="KEY_F2" to="KEY_EPG"/>
                      <key from="KEY_ENTER" to="KEY_OK"/>
              </device>
      </translate>
      <translate>
              <device name="dreambox advanced remote control (native)">
                      <key from="KEY_PLAY" to="KEY_PLAYPAUSE"/>
              </device>
      </translate>
</keymap>

* Action: Ignore keyBreak after a keyLong

After this commit we can remove all these longkeypressed stuff in python

* action: change c style cast to void* into reinterpret_cast

This commit silences the following waring:
actions/action.cpp:254:75: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]

* Fix action.cpp 'invalid use of incomplete type' error

* InputDevice: use correct IOC_SIZEBITS for non mips arch

Hardcoding the IOC_SIZEBITS is very bad because arch can override that value.
Use 13 for mips or the default 14.

Most probably it is be better to export that function from enigma2 to python.
So python will not have to "guess" the right values.

Above fixes the following error on arm:
[iInputDevices] getInputDevices event0 <ERROR: ioctl(EVIOCGNAME): [Errno 22] Invalid argument >

By Athoik
https://github.com/OpenPLi/enigma2/commit/0720302b55e5084af646e6eefef8921e9ccfa1c3

* Fix key repeat for aarch64 bit systems

* [SystemInfo] Add "RcTypeChangable"

and add an exception(not working) multi remote control for
et8500/7000/7500

Adapted from:
OpenPLi/enigma2@5b7aeb4

* [InputDevice] Re-add box branding

* InputDevice: Simplify RcTypeControl class

Do not use pathExists but use directly os.path.exists

Also, make sure that rctype has value else write the default 0, this prevents
the following error to happen when switching RC.

TypeError: %d format: a number is required, not NoneType

Adapted from:
OpenPLi/enigma2@10d3a81

* [InputDevice] Alter debug line
Most other files use the python filename

Co-authored-by: Mirakels <mirakels@openpli.org>
Co-authored-by: littlesat <littlesat99@yahoo.com>
Co-authored-by: Athanasios Oikonomou <athoik@gmail.com>
Co-authored-by: Betacentauri <betacentauri@arcor.de>
Co-authored-by: Captain <captain.onboard@web.de>